better message handling in existing assets table
